+++ .build/netcoreandfx.props
@@ -3,6 +3,7 @@
   <PropertyGroup>
     <TargetFrameworks>net47;net471;net472;net48;net481;net6.0-windows;net7.0-windows;net8.0-windows</TargetFrameworks>
     <LangVersion>latest</LangVersion>
+    <ImplicitUsings>true</ImplicitUsings>
   </PropertyGroup>
 
   <PropertyGroup>
