+++ src/ConsoleRunner/Exceptions/RandomException.cs
@@ -1,25 +1,22 @@
-﻿using System;
+﻿namespace ConsoleRunner.Exceptions;
 
-namespace ConsoleRunner.Exceptions
+/// <summary>
+/// An exception which occurs randomly.
+/// </summary>
+[Serializable]
+public class RandomException : Exception
 {
-    /// <summary>
-    /// An exception which occurs randomly.
-    /// </summary>
-    [Serializable]
-    public class RandomException : Exception
+    public RandomException()
+        : this("A random exception occcured.")
     {
-        public RandomException()
-            : this("A random exception occcured.")
-        {
-        }
+    }
 
-        public RandomException(string message) : base(message)
-        {
-        }
+    public RandomException(string message) : base(message)
+    {
+    }
 
-        public RandomException(string message, Exception innerException)
-            : base(message, innerException)
-        {
-        }
+    public RandomException(string message, Exception innerException)
+        : base(message, innerException)
+    {
     }
 }
