--- src/ConsoleRunner/Infrastructure/ContainerBindings.cs
@@ -1,15 +1,0 @@
-﻿using Autofac;
-using ConsoleRunner.Infrastructure.Diagnostics;
-
-namespace ConsoleRunner.Infrastructure
-{
-    class ContainerBindings : Module
-    {
-        protected override void Load(ContainerBuilder builder)
-        {
-            builder.RegisterType<DotRasLoggingAdapter>().AsImplementedInterfaces();
-
-            base.Load(builder);
-        }
-    }
-}
