+++ src/ConsoleRunner/Program.Main.cs
@@ -1,46 +1,48 @@
-﻿using System;
-using System.Threading;
-using System.Threading.Tasks;
+﻿using ConsoleRunner.Diagnostics;
+using DotRas.Diagnostics;
+using Microsoft.Extensions.DependencyInjection;
 
-namespace ConsoleRunner
+namespace ConsoleRunner;
+
+partial class Program
 {
-    partial class Program
+    private static readonly CancellationTokenSource CancellationSource = new CancellationTokenSource();
+
+    public static async Task Main()
     {
-        private static readonly CancellationTokenSource CancellationSource = new CancellationTokenSource();
+        LoggerLocator.SetLocator(serviceProvider.GetRequiredService<DotRasLoggingAdapter>);
 
-        public static async Task Main()
+        Console.WriteLine("Press CTRL+C at any time to to cancel the application...");
+        Console.WriteLine();
+
+        AttachCancellationSourceToCancelKeyPress();
+
+        try
         {
-            Console.WriteLine("Press CTRL+C at any time to to cancel the application...");
-            Console.WriteLine();
+            ConfigureIoC();
+            ConfigureApplication();
 
-            AttachCancellationSourceToCancelKeyPress();
-
-            try
-            {
-                ConfigureIoC();
-                ConfigureApplication();
-
-                using var program = new Program();
-                await program.RunAsync();
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine(ex);
-            }
-            finally {
-                CancellationSource.Dispose();
-            }
+            using var program = new Program();
+            await program.RunAsync();
         }
-
-        private static void AttachCancellationSourceToCancelKeyPress()
+        catch (Exception ex)
         {
-            Console.CancelKeyPress += (sender, e) =>
-            {
-                Console.WriteLine("Terminating the application...");
-
-                CancellationSource.Cancel();
-                e.Cancel = true;
-            };
+            Console.WriteLine(ex);
         }
+        finally
+        {
+            CancellationSource.Dispose();
+        }
+    }
+
+    private static void AttachCancellationSourceToCancelKeyPress()
+    {
+        Console.CancelKeyPress += (sender, e) =>
+        {
+            Console.WriteLine("Terminating the application...");
+
+            CancellationSource.Cancel();
+            e.Cancel = true;
+        };
     }
 }
